Add `calculate_shares_delta_given_bonds_delta_derivative` (#7)
# Resolved Issues
  This is a part of adding functions needed to test remove_liquidity

# Description
- Adds the lower level function
`calculate_shares_delta_given_bonds_delta_derivative` to lp.rs
- Adds a new version of `calculate_present_value_safe` that doesn't
panic. Future work will remove the previous version.
  - Adds a helper function to calculate idle reserves
  - Tests `calculate_shares_delta_given_bonds_delta_derivative`. 

I ran diagnostics to make sure that we were actually testing the
function, the test errors on the present value calculation about half
the time, errors on the actual function 1/4 time and succeeds about a
1/4 of the time.


# Review Checklists

Please check each item **before approving** the pull request. While
going
through the checklist, it is recommended to leave comments on items that
are
referenced in the checklist to make sure that they are reviewed.

- [ ] **Testing**
    - [ ] Are there new or updated unit or integration tests?
    - [ ] Do the tests cover the happy paths?
    - [ ] Do the tests cover the unhappy paths?
- [ ] Are there an adequate number of fuzz tests to ensure that we are
          covering the full input space?
- [ ] If matching Solidity behavior, are there differential fuzz tests
that
          ensure that Rust matches Solidity?