fix calculate_implied_rate to scale according to position duration (#6)
PR originally from
[delvtech/hyperdrive#1006](https://github.com/delvtech/hyperdrive/pull/1006)

# Resolved Issues
Fixes [delvtech/hyperdrive/#1003
](https://github.com/delvtech/hyperdrive/issues/1003)

# Description
Fixes calculate_implied_rate to adjust the variable rate according to
the term duration and compounding frequency.

We should probably add a test with a position duration different than 1
year, but I couldn't easily figure out how to.

# Review Checklists

Please check each item **before approving** the pull request. While
going
through the checklist, it is recommended to leave comments on items that
are
referenced in the checklist to make sure that they are reviewed. If
there are
multiple reviewers, copy the checklists into sections titled `##
[Reviewer Name]`.
If the PR doesn't touch Solidity and/or Rust, the corresponding
checklist can
be removed.

## [[Reviewer Name]]

### Rust

- [ ] **Testing**
    - [ ] Are there new or updated unit or integration tests?
    - [ ] Do the tests cover the happy paths?
    - [ ] Do the tests cover the unhappy paths?
- [ ] Are there an adequate number of fuzz tests to ensure that we are
          covering the full input space?
- [ ] If matching Solidity behavior, are there differential fuzz tests
that
          ensure that Rust matches Solidity?