Hardcode hyperdrive version dependency (#14)
The code in this repo is built against a specific version of Hyperdrive,
therefore, like all dependencies, the version should be hardcoded and
committed to main. This PR removes the script to dynamically fetch the
latest release and adds a `hyperdrive.version` file which is read from
the wrappers build script.