fix `calculate_spot_price_after_short` and add open short deltas & preview (#15)
`calculate_spot_price_after_short` did not correctly compute the share
deltas when a `base_amount` was not given. This PR adds a test to
demonstrate this as well as a fix.

Part of the process also required adding `calculate_pool_deltas_after_open_short`
and a corresponding test.

Since I was most of the way there, I also added a `calculate_pool_state_after_open_short`
that returns a mutated state after the reserve deltas have been applied.