Update hyperdrive-wrappers to hyperdrive v1.0.5 (#75)
# Resolved Issues

# Description
Commits tell a pretty good story, I'd recommend using them. This PR gets
us to hyperdrive version 1.0.5 which had some breaking changes. The big
one was fixing the deployments that required linking LPMath in the
bytecode. The wrappers all now have their own `deploy_linked_contract()`
method that can do this for us. The rest are in this list and a pretty
well separated by commit.

- update version to 1.0.5
- remove `deploy()` aliases from build.rs since 1.0.5 renames the
conflicting methods.
- add circuit breaker delta where needed
- remove target4 calls
- add `deploy_linked_contract()` method to the wrappers
- fix deployments that require LPMath linked reference.
- fix `Checkpoint` calls

# Review Checklists

Please check each item **before approving** the pull request. While
going
through the checklist, it is recommended to leave comments on items that
are
referenced in the checklist to make sure that they are reviewed.

- [ ] **Testing**
    - [ ] Are there new or updated unit or integration tests?
    - [ ] Do the tests cover the happy paths?
    - [ ] Do the tests cover the unhappy paths?
- [ ] Are there an adequate number of fuzz tests to ensure that we are
          covering the full input space?
- [ ] If matching Solidity behavior, are there differential fuzz tests
that
          ensure that Rust matches Solidity?

---------

Co-authored-by: Ryan Goree <goree.ryan@gmail.com>
Co-authored-by: Jonny Rhea <5555162+jrhea@users.noreply.github.com>
Co-authored-by: jonny rhea <jonathan.rhea@gmail.com>
Co-authored-by: Alex Towle <jalextowle@gmail.com>