Fix cyclic dependency between `test-utils` and `hyperdrive-math` crates (#92)
# Description
hyperdrive-math was depending on test-utils and vice versa. The culprit
is the `Agent` struct in `test_utils` that required a lot of things in
`hyperdrive-math` for the `Agent` methods. `hyperdrive-math` then
imported the Agent to run tests.

This PR removes the cyclic dependency by extending the `Agent` struct in
`hyperdrive-math` with it's own `HyperdriveMathAgent` in
test_utils/agent.rs. The agent module is `#[cfg(test)]` decorated so it
only compiles for tests and isn't included in the release binary.

One caveat to this approach though is that other crates will not be able
to import the HyperdriveMathAgent, including the tests/ directory, so I
had to move tests/integration_tests.rs to
src/test_utils/integration_tests.rs

# Review Checklists

Please check each item **before approving** the pull request. While
going
through the checklist, it is recommended to leave comments on items that
are
referenced in the checklist to make sure that they are reviewed.

- [ ] **Testing**
    - [ ] Are there new or updated unit or integration tests?
    - [ ] Do the tests cover the happy paths?
    - [ ] Do the tests cover the unhappy paths?
- [ ] Are there an adequate number of fuzz tests to ensure that we are
          covering the full input space?
- [ ] If matching Solidity behavior, are there differential fuzz tests
that
          ensure that Rust matches Solidity?