fee overhaul & targeted long fix (#95)
# Description
This PR is the result of an audit of the fee functions. The following
changes were made (one commit per bullet):

1. revert the rounding adjustment from #97. We realized that the bug
actually lies in the solidity _test utils_. This has been fixed so I
reverted the rounding behavior to match.
2. update docstrings to use a more consistent format. I introduce a
convention of my own to use upper-case $\Phi_{c,ol}(...)$ for the fee
function and lower-case $\phi_{c}$ for the fee multiplier constant. The
subscript on the fee function is the fee type $\in c, g, f$ for curve,
gov, and flat; as well as the trade type $\in ol, cl, os, cs$ for
opening or closing a long or short. This replaces the inconsistent
method of using e.g. `c(x)` at some times and `curve_fee` at other
times, or using repeat symbols for different equations.
3. add `maybe_curve_fee` arg to gov fee calculations to fix #43 and
reduce computation.
4. change the rounding behavior to match solidity (note it's matching
the eqs in `internal/HyperdriveBase.sol`
5. the above changes revealed a unit issue in `calculate_targeted_long`
where the gov fee (and derivative) was not converted to shares. I
updated my [overleaf
doc](https://www.overleaf.com/read/jkngspgyqzkf#1de6ce), the docstrings,
and the code itself.