Refactor crates to remove cyclic dependencies (#105)
# Resolved Issues

Related to #92

# Description

This refactors the crates to remove all cyclic dependencies. Below is an
illustration of the dependency tree before and after with a few notes:


![hyperdrive-rs-dep-trees](https://github.com/delvtech/hyperdrive-rs/assets/3289505/1ac060aa-58c9-4694-bf49-3ceff5aa2547)

# Review Checklists

Please check each item **before approving** the pull request. While
going
through the checklist, it is recommended to leave comments on items that
are
referenced in the checklist to make sure that they are reviewed.

- [ ] **Testing**
    - [ ] Are there new or updated unit or integration tests?
    - [ ] Do the tests cover the happy paths?
    - [ ] Do the tests cover the unhappy paths?
- [ ] Are there an adequate number of fuzz tests to ensure that we are
          covering the full input space?
- [ ] If matching Solidity behavior, are there differential fuzz tests
that
          ensure that Rust matches Solidity?