manually disable build (#148)
# Resolved Issues
#141 

# Description
Instead of automatically checking if files have changed, just make this
process manual via the `BUILD_DISABLED` environment variable. Most of
the time for development this can just be set to `true` if solidity
files aren't being worked on. If solidity files are being edited then
the variable can be set to `false` to enable automatic rebuilding. This
won't affect CI since it defaults to `false`.

# Review Checklists

Please check each item **before approving** the pull request. While
going
through the checklist, it is recommended to leave comments on items that
are
referenced in the checklist to make sure that they are reviewed.

- [ ] **Testing**
    - [ ] Are there new or updated unit or integration tests?
    - [ ] Do the tests cover the happy paths?
    - [ ] Do the tests cover the unhappy paths?
- [ ] Are there an adequate number of fuzz tests to ensure that we are
          covering the full input space?
- [ ] If matching Solidity behavior, are there differential fuzz tests
that
          ensure that Rust matches Solidity?