fixes for max short test (#209)
# Description
The primary contributions in this PR are:
1. update the minimum share reserves after short from `z_min + max(zeta,
0) + e/c` to `z_min + max(zeta, e/c, 0)`. This update is justified
because the checks are performed independently, and thus the reserves do
not need to account for share reserves **plus** exposure. Just whichever
is larger.
2. update the approximate absolute max short to account for the fact
that opening a short reduces exposure. Here is the updated logic:


![image](https://github.com/user-attachments/assets/b67aa8a4-1052-4729-8c74-74be98805c86)

![image](https://github.com/user-attachments/assets/5deb01cb-1acc-477a-b986-11707036ef11)

## Other changes:
- updated a handful of functions to return `FixedPoint::<I256>` instead
of `FixedPoint::<U256>`. This cleaned up the logic in the max short
functions.
- `fuzz_sol_calculate_max_long` gets a tolerance -- it was
intermittently failing in PRs. Later we should fix this, as requested in
https://github.com/delvtech/hyperdrive-rs/issues/136