cleanup tests & test support code (#211)
# Description
- I moved a validity check from `calculate_open_short` to
`solvency_after_short`. And we now call `solvency_after_short` in
`calculate_open_short`. This diverges a bit from Solidity because we do
not split the `apply_open_short` function from the
`calculate_open_short` function in rust. This means it was possible to
get a valid answer from `calculate_open_short` that would revert in the
contract. That should be fixed, now.
- I updated a bunch of tests to check if a short is possible by directly
calling `solvency_after_short(min_txn)` instead of the incorrect check
it was doing.
- I removed the `get_max_short` function from the preamble. We don't
need all of the extra checks.
- I modified the preamble random state generator to be faster by taking
fewer steps and using the approximate max short. This was a really slow
step in the solidity tests.
- the Agent class now uses the new functions when you call
`agent.calculate_max_short`
- I increased the defaults for `calculate_short_bonds_given_deposit` to
`maybe_base_tolerance=1e10` (from `1e9`) and
`maybe_max_iterations=1_000` (from `500`).

---------

Co-authored-by: Matthew Brown <mattfdbrown@gmail.com>